Model/requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.8.0
dnspython==1.16.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.12.1
gast==0.4.0
google-auth==2.25.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.2
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.11.0
langchain==0.0.348
langchain-core==0.0.12
langsmith==0.0.69
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.0
opt-einsum==3.3.0
packaging==23.2
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==3.6.0
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
